requirements.txt
praw>=7.7.0
supabase>=2.0.0
openai>=1.0.0
vaderSentiment>=3.3.2
python-dotenv>=1.0.0
flask>=3.0.0
requests>=2.31.0
